name change to avoid gem conflicts | ffi_gen -> ffi_generator
