+++ lib/ffi_generator/clang.rb
