+++ lib/ffi_generator/clang/c.rb
