+++ lib/ffi_generator/clang/index.rb
