+++ lib/ffi_generator/clang/string.rb
