+++ lib/ffi_generator/clang/token.rb
