+++ lib/ffi_generator/clang/translation_unit.rb
