+++ lib/ffi_generator/generator/array_type.rb
