+++ lib/ffi_generator/generator/by_value_type.rb
