+++ lib/ffi_generator/generator/define.rb
