+++ lib/ffi_generator/generator/enum.rb
