+++ lib/ffi_generator/generator/java.rb
