+++ lib/ffi_generator/generator/java_pre.java
