+++ lib/ffi_generator/generator/name.rb
