+++ lib/ffi_generator/generator/pointer_type.rb
