+++ lib/ffi_generator/generator/primitive_type.rb
