+++ lib/ffi_generator/generator/ruby.rb
