+++ lib/ffi_generator/generator/string_type.rb
