+++ lib/ffi_generator/generator/struct_or_union.rb
