+++ lib/ffi_generator/generator/writer.rb
