+++ test/output/GL/freeglut_std.rb
@@ -5,1275 +5,1275 @@ require 'ffi'
 module GLUT
   extend FFI::Library
   ffi_lib 'glut'
-  
+
   def self.attach_function(name, *_)
     begin; super; rescue FFI::NotFoundError => e
       (class << self; self; end).class_eval { define_method(name) { |*_| raise e } }
     end
   end
-  
+
   FREEGLUT = 1
-  
+
   API_VERSION = 4
-  
+
   FREEGLUT_VERSION_2_0 = 1
-  
+
   XLIB_IMPLEMENTATION = 13
-  
+
   KEY_F1 = 0x0001
-  
+
   KEY_F2 = 0x0002
-  
+
   KEY_F3 = 0x0003
-  
+
   KEY_F4 = 0x0004
-  
+
   KEY_F5 = 0x0005
-  
+
   KEY_F6 = 0x0006
-  
+
   KEY_F7 = 0x0007
-  
+
   KEY_F8 = 0x0008
-  
+
   KEY_F9 = 0x0009
-  
+
   KEY_F10 = 0x000A
-  
+
   KEY_F11 = 0x000B
-  
+
   KEY_F12 = 0x000C
-  
+
   KEY_LEFT = 0x0064
-  
+
   KEY_UP = 0x0065
-  
+
   KEY_RIGHT = 0x0066
-  
+
   KEY_DOWN = 0x0067
-  
+
   KEY_PAGE_UP = 0x0068
-  
+
   KEY_PAGE_DOWN = 0x0069
-  
+
   KEY_HOME = 0x006A
-  
+
   KEY_END = 0x006B
-  
+
   KEY_INSERT = 0x006C
-  
+
   LEFT_BUTTON = 0x0000
-  
+
   MIDDLE_BUTTON = 0x0001
-  
+
   RIGHT_BUTTON = 0x0002
-  
+
   DOWN = 0x0000
-  
+
   UP = 0x0001
-  
+
   LEFT = 0x0000
-  
+
   ENTERED = 0x0001
-  
+
   RGB = 0x0000
-  
+
   RGBA = 0x0000
-  
+
   INDEX = 0x0001
-  
+
   SINGLE = 0x0000
-  
+
   DOUBLE = 0x0002
-  
+
   ACCUM = 0x0004
-  
+
   ALPHA = 0x0008
-  
+
   DEPTH = 0x0010
-  
+
   STENCIL = 0x0020
-  
+
   MULTISAMPLE = 0x0080
-  
+
   STEREO = 0x0100
-  
+
   LUMINANCE = 0x0200
-  
+
   MENU_NOT_IN_USE = 0x0000
-  
+
   MENU_IN_USE = 0x0001
-  
+
   NOT_VISIBLE = 0x0000
-  
+
   VISIBLE = 0x0001
-  
+
   HIDDEN = 0x0000
-  
+
   FULLY_RETAINED = 0x0001
-  
+
   PARTIALLY_RETAINED = 0x0002
-  
+
   FULLY_COVERED = 0x0003
-  
+
   WINDOW_X = 0x0064
-  
+
   WINDOW_Y = 0x0065
-  
+
   WINDOW_WIDTH = 0x0066
-  
+
   WINDOW_HEIGHT = 0x0067
-  
+
   WINDOW_BUFFER_SIZE = 0x0068
-  
+
   WINDOW_STENCIL_SIZE = 0x0069
-  
+
   WINDOW_DEPTH_SIZE = 0x006A
-  
+
   WINDOW_RED_SIZE = 0x006B
-  
+
   WINDOW_GREEN_SIZE = 0x006C
-  
+
   WINDOW_BLUE_SIZE = 0x006D
-  
+
   WINDOW_ALPHA_SIZE = 0x006E
-  
+
   WINDOW_ACCUM_RED_SIZE = 0x006F
-  
+
   WINDOW_ACCUM_GREEN_SIZE = 0x0070
-  
+
   WINDOW_ACCUM_BLUE_SIZE = 0x0071
-  
+
   WINDOW_ACCUM_ALPHA_SIZE = 0x0072
-  
+
   WINDOW_DOUBLEBUFFER = 0x0073
-  
+
   WINDOW_RGBA = 0x0074
-  
+
   WINDOW_PARENT = 0x0075
-  
+
   WINDOW_NUM_CHILDREN = 0x0076
-  
+
   WINDOW_COLORMAP_SIZE = 0x0077
-  
+
   WINDOW_NUM_SAMPLES = 0x0078
-  
+
   WINDOW_STEREO = 0x0079
-  
+
   WINDOW_CURSOR = 0x007A
-  
+
   SCREEN_WIDTH = 0x00C8
-  
+
   SCREEN_HEIGHT = 0x00C9
-  
+
   SCREEN_WIDTH_MM = 0x00CA
-  
+
   SCREEN_HEIGHT_MM = 0x00CB
-  
+
   MENU_NUM_ITEMS = 0x012C
-  
+
   DISPLAY_MODE_POSSIBLE = 0x0190
-  
+
   INIT_WINDOW_X = 0x01F4
-  
+
   INIT_WINDOW_Y = 0x01F5
-  
+
   INIT_WINDOW_WIDTH = 0x01F6
-  
+
   INIT_WINDOW_HEIGHT = 0x01F7
-  
+
   INIT_DISPLAY_MODE = 0x01F8
-  
+
   ELAPSED_TIME = 0x02BC
-  
+
   WINDOW_FORMAT_ID = 0x007B
-  
+
   HAS_KEYBOARD = 0x0258
-  
+
   HAS_MOUSE = 0x0259
-  
+
   HAS_SPACEBALL = 0x025A
-  
+
   HAS_DIAL_AND_BUTTON_BOX = 0x025B
-  
+
   HAS_TABLET = 0x025C
-  
+
   NUM_MOUSE_BUTTONS = 0x025D
-  
+
   NUM_SPACEBALL_BUTTONS = 0x025E
-  
+
   NUM_BUTTON_BOX_BUTTONS = 0x025F
-  
+
   NUM_DIALS = 0x0260
-  
+
   NUM_TABLET_BUTTONS = 0x0261
-  
+
   DEVICE_IGNORE_KEY_REPEAT = 0x0262
-  
+
   DEVICE_KEY_REPEAT = 0x0263
-  
+
   HAS_JOYSTICK = 0x0264
-  
+
   OWNS_JOYSTICK = 0x0265
-  
+
   JOYSTICK_BUTTONS = 0x0266
-  
+
   JOYSTICK_AXES = 0x0267
-  
+
   JOYSTICK_POLL_RATE = 0x0268
-  
+
   OVERLAY_POSSIBLE = 0x0320
-  
+
   LAYER_IN_USE = 0x0321
-  
+
   HAS_OVERLAY = 0x0322
-  
+
   TRANSPARENT_INDEX = 0x0323
-  
+
   NORMAL_DAMAGED = 0x0324
-  
+
   OVERLAY_DAMAGED = 0x0325
-  
+
   VIDEO_RESIZE_POSSIBLE = 0x0384
-  
+
   VIDEO_RESIZE_IN_USE = 0x0385
-  
+
   VIDEO_RESIZE_X_DELTA = 0x0386
-  
+
   VIDEO_RESIZE_Y_DELTA = 0x0387
-  
+
   VIDEO_RESIZE_WIDTH_DELTA = 0x0388
-  
+
   VIDEO_RESIZE_HEIGHT_DELTA = 0x0389
-  
+
   VIDEO_RESIZE_X = 0x038A
-  
+
   VIDEO_RESIZE_Y = 0x038B
-  
+
   VIDEO_RESIZE_WIDTH = 0x038C
-  
+
   VIDEO_RESIZE_HEIGHT = 0x038D
-  
+
   NORMAL = 0x0000
-  
+
   OVERLAY = 0x0001
-  
+
   ACTIVE_SHIFT = 0x0001
-  
+
   ACTIVE_CTRL = 0x0002
-  
+
   ACTIVE_ALT = 0x0004
-  
+
   CURSOR_RIGHT_ARROW = 0x0000
-  
+
   CURSOR_LEFT_ARROW = 0x0001
-  
+
   CURSOR_INFO = 0x0002
-  
+
   CURSOR_DESTROY = 0x0003
-  
+
   CURSOR_HELP = 0x0004
-  
+
   CURSOR_CYCLE = 0x0005
-  
+
   CURSOR_SPRAY = 0x0006
-  
+
   CURSOR_WAIT = 0x0007
-  
+
   CURSOR_TEXT = 0x0008
-  
+
   CURSOR_CROSSHAIR = 0x0009
-  
+
   CURSOR_UP_DOWN = 0x000A
-  
+
   CURSOR_LEFT_RIGHT = 0x000B
-  
+
   CURSOR_TOP_SIDE = 0x000C
-  
+
   CURSOR_BOTTOM_SIDE = 0x000D
-  
+
   CURSOR_LEFT_SIDE = 0x000E
-  
+
   CURSOR_RIGHT_SIDE = 0x000F
-  
+
   CURSOR_TOP_LEFT_CORNER = 0x0010
-  
+
   CURSOR_TOP_RIGHT_CORNER = 0x0011
-  
+
   CURSOR_BOTTOM_RIGHT_CORNER = 0x0012
-  
+
   CURSOR_BOTTOM_LEFT_CORNER = 0x0013
-  
+
   CURSOR_INHERIT = 0x0064
-  
+
   CURSOR_NONE = 0x0065
-  
+
   CURSOR_FULL_CROSSHAIR = 0x0066
-  
+
   RED = 0x0000
-  
+
   GREEN = 0x0001
-  
+
   BLUE = 0x0002
-  
+
   KEY_REPEAT_OFF = 0x0000
-  
+
   KEY_REPEAT_ON = 0x0001
-  
+
   KEY_REPEAT_DEFAULT = 0x0002
-  
+
   JOYSTICK_BUTTON_A = 0x0001
-  
+
   JOYSTICK_BUTTON_B = 0x0002
-  
+
   JOYSTICK_BUTTON_C = 0x0004
-  
+
   JOYSTICK_BUTTON_D = 0x0008
-  
+
   GAME_MODE_ACTIVE = 0x0000
-  
+
   GAME_MODE_POSSIBLE = 0x0001
-  
+
   GAME_MODE_WIDTH = 0x0002
-  
+
   GAME_MODE_HEIGHT = 0x0003
-  
+
   GAME_MODE_PIXEL_DEPTH = 0x0004
-  
+
   GAME_MODE_REFRESH_RATE = 0x0005
-  
+
   GAME_MODE_DISPLAY_CHANGED = 0x0006
-  
+
   # Initialization functions, see fglut_init.c
-  # 
+  #
   # @method init(pargc, argv)
-  # @param [FFI::Pointer(*Int)] pargc 
-  # @param [FFI::Pointer(**CharS)] argv 
-  # @return [nil] 
+  # @param [FFI::Pointer(*Int)] pargc
+  # @param [FFI::Pointer(**CharS)] argv
+  # @return [nil]
   # @scope class
   attach_function :init, :glutInit, [:pointer, :pointer], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method init_window_position(x, y)
-  # @param [Integer] x 
-  # @param [Integer] y 
-  # @return [nil] 
+  # @param [Integer] x
+  # @param [Integer] y
+  # @return [nil]
   # @scope class
   attach_function :init_window_position, :glutInitWindowPosition, [:int, :int], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method init_window_size(width, height)
-  # @param [Integer] width 
-  # @param [Integer] height 
-  # @return [nil] 
+  # @param [Integer] width
+  # @param [Integer] height
+  # @return [nil]
   # @scope class
   attach_function :init_window_size, :glutInitWindowSize, [:int, :int], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method init_display_mode(display_mode)
-  # @param [Integer] display_mode 
-  # @return [nil] 
+  # @param [Integer] display_mode
+  # @return [nil]
   # @scope class
   attach_function :init_display_mode, :glutInitDisplayMode, [:uint], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method init_display_string(display_mode)
-  # @param [String] display_mode 
-  # @return [nil] 
+  # @param [String] display_mode
+  # @return [nil]
   # @scope class
   attach_function :init_display_string, :glutInitDisplayString, [:string], :void
-  
+
   # Process loop function, see freeglut_main.c
-  # 
+  #
   # @method main_loop()
-  # @return [nil] 
+  # @return [nil]
   # @scope class
   attach_function :main_loop, :glutMainLoop, [], :void
-  
+
   # Window management functions, see freeglut_window.c
-  # 
+  #
   # @method create_window(title)
-  # @param [String] title 
-  # @return [Integer] 
+  # @param [String] title
+  # @return [Integer]
   # @scope class
   attach_function :create_window, :glutCreateWindow, [:string], :int
-  
+
   # (Not documented)
-  # 
+  #
   # @method create_sub_window(window, x, y, width, height)
-  # @param [Integer] window 
-  # @param [Integer] x 
-  # @param [Integer] y 
-  # @param [Integer] width 
-  # @param [Integer] height 
-  # @return [Integer] 
+  # @param [Integer] window
+  # @param [Integer] x
+  # @param [Integer] y
+  # @param [Integer] width
+  # @param [Integer] height
+  # @return [Integer]
   # @scope class
   attach_function :create_sub_window, :glutCreateSubWindow, [:int, :int, :int, :int, :int], :int
-  
+
   # (Not documented)
-  # 
+  #
   # @method destroy_window(window)
-  # @param [Integer] window 
-  # @return [nil] 
+  # @param [Integer] window
+  # @return [nil]
   # @scope class
   attach_function :destroy_window, :glutDestroyWindow, [:int], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method set_window(window)
-  # @param [Integer] window 
-  # @return [nil] 
+  # @param [Integer] window
+  # @return [nil]
   # @scope class
   attach_function :set_window, :glutSetWindow, [:int], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method get_window()
-  # @return [Integer] 
+  # @return [Integer]
   # @scope class
   attach_function :get_window, :glutGetWindow, [], :int
-  
+
   # (Not documented)
-  # 
+  #
   # @method set_window_title(title)
-  # @param [String] title 
-  # @return [nil] 
+  # @param [String] title
+  # @return [nil]
   # @scope class
   attach_function :set_window_title, :glutSetWindowTitle, [:string], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method set_icon_title(title)
-  # @param [String] title 
-  # @return [nil] 
+  # @param [String] title
+  # @return [nil]
   # @scope class
   attach_function :set_icon_title, :glutSetIconTitle, [:string], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method reshape_window(width, height)
-  # @param [Integer] width 
-  # @param [Integer] height 
-  # @return [nil] 
+  # @param [Integer] width
+  # @param [Integer] height
+  # @return [nil]
   # @scope class
   attach_function :reshape_window, :glutReshapeWindow, [:int, :int], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method position_window(x, y)
-  # @param [Integer] x 
-  # @param [Integer] y 
-  # @return [nil] 
+  # @param [Integer] x
+  # @param [Integer] y
+  # @return [nil]
   # @scope class
   attach_function :position_window, :glutPositionWindow, [:int, :int], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method show_window()
-  # @return [nil] 
+  # @return [nil]
   # @scope class
   attach_function :show_window, :glutShowWindow, [], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method hide_window()
-  # @return [nil] 
+  # @return [nil]
   # @scope class
   attach_function :hide_window, :glutHideWindow, [], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method iconify_window()
-  # @return [nil] 
+  # @return [nil]
   # @scope class
   attach_function :iconify_window, :glutIconifyWindow, [], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method push_window()
-  # @return [nil] 
+  # @return [nil]
   # @scope class
   attach_function :push_window, :glutPushWindow, [], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method pop_window()
-  # @return [nil] 
+  # @return [nil]
   # @scope class
   attach_function :pop_window, :glutPopWindow, [], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method full_screen()
-  # @return [nil] 
+  # @return [nil]
   # @scope class
   attach_function :full_screen, :glutFullScreen, [], :void
-  
+
   # Display-connected functions, see freeglut_display.c
-  # 
+  #
   # @method post_window_redisplay(window)
-  # @param [Integer] window 
-  # @return [nil] 
+  # @param [Integer] window
+  # @return [nil]
   # @scope class
   attach_function :post_window_redisplay, :glutPostWindowRedisplay, [:int], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method post_redisplay()
-  # @return [nil] 
+  # @return [nil]
   # @scope class
   attach_function :post_redisplay, :glutPostRedisplay, [], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method swap_buffers()
-  # @return [nil] 
+  # @return [nil]
   # @scope class
   attach_function :swap_buffers, :glutSwapBuffers, [], :void
-  
+
   # Mouse cursor functions, see freeglut_cursor.c
-  # 
+  #
   # @method warp_pointer(x, y)
-  # @param [Integer] x 
-  # @param [Integer] y 
-  # @return [nil] 
+  # @param [Integer] x
+  # @param [Integer] y
+  # @return [nil]
   # @scope class
   attach_function :warp_pointer, :glutWarpPointer, [:int, :int], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method set_cursor(cursor)
-  # @param [Integer] cursor 
-  # @return [nil] 
+  # @param [Integer] cursor
+  # @return [nil]
   # @scope class
   attach_function :set_cursor, :glutSetCursor, [:int], :void
-  
+
   # Overlay stuff, see freeglut_overlay.c
-  # 
+  #
   # @method establish_overlay()
-  # @return [nil] 
+  # @return [nil]
   # @scope class
   attach_function :establish_overlay, :glutEstablishOverlay, [], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method remove_overlay()
-  # @return [nil] 
+  # @return [nil]
   # @scope class
   attach_function :remove_overlay, :glutRemoveOverlay, [], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method use_layer(layer)
-  # @param [Integer] layer 
-  # @return [nil] 
+  # @param [Integer] layer
+  # @return [nil]
   # @scope class
   attach_function :use_layer, :glutUseLayer, [:uint], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method post_overlay_redisplay()
-  # @return [nil] 
+  # @return [nil]
   # @scope class
   attach_function :post_overlay_redisplay, :glutPostOverlayRedisplay, [], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method post_window_overlay_redisplay(window)
-  # @param [Integer] window 
-  # @return [nil] 
+  # @param [Integer] window
+  # @return [nil]
   # @scope class
   attach_function :post_window_overlay_redisplay, :glutPostWindowOverlayRedisplay, [:int], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method show_overlay()
-  # @return [nil] 
+  # @return [nil]
   # @scope class
   attach_function :show_overlay, :glutShowOverlay, [], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method hide_overlay()
-  # @return [nil] 
+  # @return [nil]
   # @scope class
   attach_function :hide_overlay, :glutHideOverlay, [], :void
-  
+
   # Menu stuff, see freeglut_menu.c
-  # 
+  #
   # @method create_menu(callback)
-  # @param [FFI::Pointer(*)] callback 
-  # @return [Integer] 
+  # @param [FFI::Pointer(*)] callback
+  # @return [Integer]
   # @scope class
   attach_function :create_menu, :glutCreateMenu, [:pointer], :int
-  
+
   # (Not documented)
-  # 
+  #
   # @method destroy_menu(menu)
-  # @param [Integer] menu 
-  # @return [nil] 
+  # @param [Integer] menu
+  # @return [nil]
   # @scope class
   attach_function :destroy_menu, :glutDestroyMenu, [:int], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method get_menu()
-  # @return [Integer] 
+  # @return [Integer]
   # @scope class
   attach_function :get_menu, :glutGetMenu, [], :int
-  
+
   # (Not documented)
-  # 
+  #
   # @method set_menu(menu)
-  # @param [Integer] menu 
-  # @return [nil] 
+  # @param [Integer] menu
+  # @return [nil]
   # @scope class
   attach_function :set_menu, :glutSetMenu, [:int], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method add_menu_entry(label, value)
-  # @param [String] label 
-  # @param [Integer] value 
-  # @return [nil] 
+  # @param [String] label
+  # @param [Integer] value
+  # @return [nil]
   # @scope class
   attach_function :add_menu_entry, :glutAddMenuEntry, [:string, :int], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method add_sub_menu(label, sub_menu)
-  # @param [String] label 
-  # @param [Integer] sub_menu 
-  # @return [nil] 
+  # @param [String] label
+  # @param [Integer] sub_menu
+  # @return [nil]
   # @scope class
   attach_function :add_sub_menu, :glutAddSubMenu, [:string, :int], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method change_to_menu_entry(item, label, value)
-  # @param [Integer] item 
-  # @param [String] label 
-  # @param [Integer] value 
-  # @return [nil] 
+  # @param [Integer] item
+  # @param [String] label
+  # @param [Integer] value
+  # @return [nil]
   # @scope class
   attach_function :change_to_menu_entry, :glutChangeToMenuEntry, [:int, :string, :int], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method change_to_sub_menu(item, label, value)
-  # @param [Integer] item 
-  # @param [String] label 
-  # @param [Integer] value 
-  # @return [nil] 
+  # @param [Integer] item
+  # @param [String] label
+  # @param [Integer] value
+  # @return [nil]
   # @scope class
   attach_function :change_to_sub_menu, :glutChangeToSubMenu, [:int, :string, :int], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method remove_menu_item(item)
-  # @param [Integer] item 
-  # @return [nil] 
+  # @param [Integer] item
+  # @return [nil]
   # @scope class
   attach_function :remove_menu_item, :glutRemoveMenuItem, [:int], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method attach_menu(button)
-  # @param [Integer] button 
-  # @return [nil] 
+  # @param [Integer] button
+  # @return [nil]
   # @scope class
   attach_function :attach_menu, :glutAttachMenu, [:int], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method detach_menu(button)
-  # @param [Integer] button 
-  # @return [nil] 
+  # @param [Integer] button
+  # @return [nil]
   # @scope class
   attach_function :detach_menu, :glutDetachMenu, [:int], :void
-  
+
   # Global callback functions, see freeglut_callbacks.c
-  # 
+  #
   # @method timer_func(time, callback, value)
-  # @param [Integer] time 
-  # @param [FFI::Pointer(*)] callback 
-  # @param [Integer] value 
-  # @return [nil] 
+  # @param [Integer] time
+  # @param [FFI::Pointer(*)] callback
+  # @param [Integer] value
+  # @return [nil]
   # @scope class
   attach_function :timer_func, :glutTimerFunc, [:uint, :pointer, :int], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method idle_func(callback)
-  # @param [FFI::Pointer(*)] callback 
-  # @return [nil] 
+  # @param [FFI::Pointer(*)] callback
+  # @return [nil]
   # @scope class
   attach_function :idle_func, :glutIdleFunc, [:pointer], :void
-  
+
   # Window-specific callback functions, see freeglut_callbacks.c
-  # 
+  #
   # @method keyboard_func(callback)
-  # @param [FFI::Pointer(*)] callback 
-  # @return [nil] 
+  # @param [FFI::Pointer(*)] callback
+  # @return [nil]
   # @scope class
   attach_function :keyboard_func, :glutKeyboardFunc, [:pointer], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method special_func(callback)
-  # @param [FFI::Pointer(*)] callback 
-  # @return [nil] 
+  # @param [FFI::Pointer(*)] callback
+  # @return [nil]
   # @scope class
   attach_function :special_func, :glutSpecialFunc, [:pointer], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method reshape_func(callback)
-  # @param [FFI::Pointer(*)] callback 
-  # @return [nil] 
+  # @param [FFI::Pointer(*)] callback
+  # @return [nil]
   # @scope class
   attach_function :reshape_func, :glutReshapeFunc, [:pointer], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method visibility_func(callback)
-  # @param [FFI::Pointer(*)] callback 
-  # @return [nil] 
+  # @param [FFI::Pointer(*)] callback
+  # @return [nil]
   # @scope class
   attach_function :visibility_func, :glutVisibilityFunc, [:pointer], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method display_func(callback)
-  # @param [FFI::Pointer(*)] callback 
-  # @return [nil] 
+  # @param [FFI::Pointer(*)] callback
+  # @return [nil]
   # @scope class
   attach_function :display_func, :glutDisplayFunc, [:pointer], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method mouse_func(callback)
-  # @param [FFI::Pointer(*)] callback 
-  # @return [nil] 
+  # @param [FFI::Pointer(*)] callback
+  # @return [nil]
   # @scope class
   attach_function :mouse_func, :glutMouseFunc, [:pointer], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method motion_func(callback)
-  # @param [FFI::Pointer(*)] callback 
-  # @return [nil] 
+  # @param [FFI::Pointer(*)] callback
+  # @return [nil]
   # @scope class
   attach_function :motion_func, :glutMotionFunc, [:pointer], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method passive_motion_func(callback)
-  # @param [FFI::Pointer(*)] callback 
-  # @return [nil] 
+  # @param [FFI::Pointer(*)] callback
+  # @return [nil]
   # @scope class
   attach_function :passive_motion_func, :glutPassiveMotionFunc, [:pointer], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method entry_func(callback)
-  # @param [FFI::Pointer(*)] callback 
-  # @return [nil] 
+  # @param [FFI::Pointer(*)] callback
+  # @return [nil]
   # @scope class
   attach_function :entry_func, :glutEntryFunc, [:pointer], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method keyboard_up_func(callback)
-  # @param [FFI::Pointer(*)] callback 
-  # @return [nil] 
+  # @param [FFI::Pointer(*)] callback
+  # @return [nil]
   # @scope class
   attach_function :keyboard_up_func, :glutKeyboardUpFunc, [:pointer], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method special_up_func(callback)
-  # @param [FFI::Pointer(*)] callback 
-  # @return [nil] 
+  # @param [FFI::Pointer(*)] callback
+  # @return [nil]
   # @scope class
   attach_function :special_up_func, :glutSpecialUpFunc, [:pointer], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method joystick_func(callback, poll_interval)
-  # @param [FFI::Pointer(*)] callback 
-  # @param [Integer] poll_interval 
-  # @return [nil] 
+  # @param [FFI::Pointer(*)] callback
+  # @param [Integer] poll_interval
+  # @return [nil]
   # @scope class
   attach_function :joystick_func, :glutJoystickFunc, [:pointer, :int], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method menu_state_func(callback)
-  # @param [FFI::Pointer(*)] callback 
-  # @return [nil] 
+  # @param [FFI::Pointer(*)] callback
+  # @return [nil]
   # @scope class
   attach_function :menu_state_func, :glutMenuStateFunc, [:pointer], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method menu_status_func(callback)
-  # @param [FFI::Pointer(*)] callback 
-  # @return [nil] 
+  # @param [FFI::Pointer(*)] callback
+  # @return [nil]
   # @scope class
   attach_function :menu_status_func, :glutMenuStatusFunc, [:pointer], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method overlay_display_func(callback)
-  # @param [FFI::Pointer(*)] callback 
-  # @return [nil] 
+  # @param [FFI::Pointer(*)] callback
+  # @return [nil]
   # @scope class
   attach_function :overlay_display_func, :glutOverlayDisplayFunc, [:pointer], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method window_status_func(callback)
-  # @param [FFI::Pointer(*)] callback 
-  # @return [nil] 
+  # @param [FFI::Pointer(*)] callback
+  # @return [nil]
   # @scope class
   attach_function :window_status_func, :glutWindowStatusFunc, [:pointer], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method spaceball_motion_func(callback)
-  # @param [FFI::Pointer(*)] callback 
-  # @return [nil] 
+  # @param [FFI::Pointer(*)] callback
+  # @return [nil]
   # @scope class
   attach_function :spaceball_motion_func, :glutSpaceballMotionFunc, [:pointer], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method spaceball_rotate_func(callback)
-  # @param [FFI::Pointer(*)] callback 
-  # @return [nil] 
+  # @param [FFI::Pointer(*)] callback
+  # @return [nil]
   # @scope class
   attach_function :spaceball_rotate_func, :glutSpaceballRotateFunc, [:pointer], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method spaceball_button_func(callback)
-  # @param [FFI::Pointer(*)] callback 
-  # @return [nil] 
+  # @param [FFI::Pointer(*)] callback
+  # @return [nil]
   # @scope class
   attach_function :spaceball_button_func, :glutSpaceballButtonFunc, [:pointer], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method button_box_func(callback)
-  # @param [FFI::Pointer(*)] callback 
-  # @return [nil] 
+  # @param [FFI::Pointer(*)] callback
+  # @return [nil]
   # @scope class
   attach_function :button_box_func, :glutButtonBoxFunc, [:pointer], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method dials_func(callback)
-  # @param [FFI::Pointer(*)] callback 
-  # @return [nil] 
+  # @param [FFI::Pointer(*)] callback
+  # @return [nil]
   # @scope class
   attach_function :dials_func, :glutDialsFunc, [:pointer], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method tablet_motion_func(callback)
-  # @param [FFI::Pointer(*)] callback 
-  # @return [nil] 
+  # @param [FFI::Pointer(*)] callback
+  # @return [nil]
   # @scope class
   attach_function :tablet_motion_func, :glutTabletMotionFunc, [:pointer], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method tablet_button_func(callback)
-  # @param [FFI::Pointer(*)] callback 
-  # @return [nil] 
+  # @param [FFI::Pointer(*)] callback
+  # @return [nil]
   # @scope class
   attach_function :tablet_button_func, :glutTabletButtonFunc, [:pointer], :void
-  
+
   # State setting and retrieval functions, see freeglut_state.c
-  # 
+  #
   # @method get(query)
-  # @param [Integer] query 
-  # @return [Integer] 
+  # @param [Integer] query
+  # @return [Integer]
   # @scope class
   attach_function :get, :glutGet, [:uint], :int
-  
+
   # (Not documented)
-  # 
+  #
   # @method device_get(query)
-  # @param [Integer] query 
-  # @return [Integer] 
+  # @param [Integer] query
+  # @return [Integer]
   # @scope class
   attach_function :device_get, :glutDeviceGet, [:uint], :int
-  
+
   # (Not documented)
-  # 
+  #
   # @method get_modifiers()
-  # @return [Integer] 
+  # @return [Integer]
   # @scope class
   attach_function :get_modifiers, :glutGetModifiers, [], :int
-  
+
   # (Not documented)
-  # 
+  #
   # @method layer_get(query)
-  # @param [Integer] query 
-  # @return [Integer] 
+  # @param [Integer] query
+  # @return [Integer]
   # @scope class
   attach_function :layer_get, :glutLayerGet, [:uint], :int
-  
+
   # Font stuff, see freeglut_font.c
-  # 
+  #
   # @method bitmap_character(font, character)
-  # @param [FFI::Pointer(*Void)] font 
-  # @param [Integer] character 
-  # @return [nil] 
+  # @param [FFI::Pointer(*Void)] font
+  # @param [Integer] character
+  # @return [nil]
   # @scope class
   attach_function :bitmap_character, :glutBitmapCharacter, [:pointer, :int], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method bitmap_width(font, character)
-  # @param [FFI::Pointer(*Void)] font 
-  # @param [Integer] character 
-  # @return [Integer] 
+  # @param [FFI::Pointer(*Void)] font
+  # @param [Integer] character
+  # @return [Integer]
   # @scope class
   attach_function :bitmap_width, :glutBitmapWidth, [:pointer, :int], :int
-  
+
   # (Not documented)
-  # 
+  #
   # @method stroke_character(font, character)
-  # @param [FFI::Pointer(*Void)] font 
-  # @param [Integer] character 
-  # @return [nil] 
+  # @param [FFI::Pointer(*Void)] font
+  # @param [Integer] character
+  # @return [nil]
   # @scope class
   attach_function :stroke_character, :glutStrokeCharacter, [:pointer, :int], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method stroke_width(font, character)
-  # @param [FFI::Pointer(*Void)] font 
-  # @param [Integer] character 
-  # @return [Integer] 
+  # @param [FFI::Pointer(*Void)] font
+  # @param [Integer] character
+  # @return [Integer]
   # @scope class
   attach_function :stroke_width, :glutStrokeWidth, [:pointer, :int], :int
-  
+
   # (Not documented)
-  # 
+  #
   # @method bitmap_length(font, string)
-  # @param [FFI::Pointer(*Void)] font 
-  # @param [FFI::Pointer(*UChar)] string 
-  # @return [Integer] 
+  # @param [FFI::Pointer(*Void)] font
+  # @param [FFI::Pointer(*UChar)] string
+  # @return [Integer]
   # @scope class
   attach_function :bitmap_length, :glutBitmapLength, [:pointer, :pointer], :int
-  
+
   # (Not documented)
-  # 
+  #
   # @method stroke_length(font, string)
-  # @param [FFI::Pointer(*Void)] font 
-  # @param [FFI::Pointer(*UChar)] string 
-  # @return [Integer] 
+  # @param [FFI::Pointer(*Void)] font
+  # @param [FFI::Pointer(*UChar)] string
+  # @return [Integer]
   # @scope class
   attach_function :stroke_length, :glutStrokeLength, [:pointer, :pointer], :int
-  
+
   # Geometry functions, see freeglut_geometry.c
-  # 
+  #
   # @method wire_cube(size)
-  # @param [Float] size 
-  # @return [nil] 
+  # @param [Float] size
+  # @return [nil]
   # @scope class
   attach_function :wire_cube, :glutWireCube, [:double], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method solid_cube(size)
-  # @param [Float] size 
-  # @return [nil] 
+  # @param [Float] size
+  # @return [nil]
   # @scope class
   attach_function :solid_cube, :glutSolidCube, [:double], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method wire_sphere(radius, slices, stacks)
-  # @param [Float] radius 
-  # @param [Integer] slices 
-  # @param [Integer] stacks 
-  # @return [nil] 
+  # @param [Float] radius
+  # @param [Integer] slices
+  # @param [Integer] stacks
+  # @return [nil]
   # @scope class
   attach_function :wire_sphere, :glutWireSphere, [:double, :int, :int], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method solid_sphere(radius, slices, stacks)
-  # @param [Float] radius 
-  # @param [Integer] slices 
-  # @param [Integer] stacks 
-  # @return [nil] 
+  # @param [Float] radius
+  # @param [Integer] slices
+  # @param [Integer] stacks
+  # @return [nil]
   # @scope class
   attach_function :solid_sphere, :glutSolidSphere, [:double, :int, :int], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method wire_cone(base, height, slices, stacks)
-  # @param [Float] base 
-  # @param [Float] height 
-  # @param [Integer] slices 
-  # @param [Integer] stacks 
-  # @return [nil] 
+  # @param [Float] base
+  # @param [Float] height
+  # @param [Integer] slices
+  # @param [Integer] stacks
+  # @return [nil]
   # @scope class
   attach_function :wire_cone, :glutWireCone, [:double, :double, :int, :int], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method solid_cone(base, height, slices, stacks)
-  # @param [Float] base 
-  # @param [Float] height 
-  # @param [Integer] slices 
-  # @param [Integer] stacks 
-  # @return [nil] 
+  # @param [Float] base
+  # @param [Float] height
+  # @param [Integer] slices
+  # @param [Integer] stacks
+  # @return [nil]
   # @scope class
   attach_function :solid_cone, :glutSolidCone, [:double, :double, :int, :int], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method wire_torus(inner_radius, outer_radius, sides, rings)
-  # @param [Float] inner_radius 
-  # @param [Float] outer_radius 
-  # @param [Integer] sides 
-  # @param [Integer] rings 
-  # @return [nil] 
+  # @param [Float] inner_radius
+  # @param [Float] outer_radius
+  # @param [Integer] sides
+  # @param [Integer] rings
+  # @return [nil]
   # @scope class
   attach_function :wire_torus, :glutWireTorus, [:double, :double, :int, :int], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method solid_torus(inner_radius, outer_radius, sides, rings)
-  # @param [Float] inner_radius 
-  # @param [Float] outer_radius 
-  # @param [Integer] sides 
-  # @param [Integer] rings 
-  # @return [nil] 
+  # @param [Float] inner_radius
+  # @param [Float] outer_radius
+  # @param [Integer] sides
+  # @param [Integer] rings
+  # @return [nil]
   # @scope class
   attach_function :solid_torus, :glutSolidTorus, [:double, :double, :int, :int], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method wire_dodecahedron()
-  # @return [nil] 
+  # @return [nil]
   # @scope class
   attach_function :wire_dodecahedron, :glutWireDodecahedron, [], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method solid_dodecahedron()
-  # @return [nil] 
+  # @return [nil]
   # @scope class
   attach_function :solid_dodecahedron, :glutSolidDodecahedron, [], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method wire_octahedron()
-  # @return [nil] 
+  # @return [nil]
   # @scope class
   attach_function :wire_octahedron, :glutWireOctahedron, [], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method solid_octahedron()
-  # @return [nil] 
+  # @return [nil]
   # @scope class
   attach_function :solid_octahedron, :glutSolidOctahedron, [], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method wire_tetrahedron()
-  # @return [nil] 
+  # @return [nil]
   # @scope class
   attach_function :wire_tetrahedron, :glutWireTetrahedron, [], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method solid_tetrahedron()
-  # @return [nil] 
+  # @return [nil]
   # @scope class
   attach_function :solid_tetrahedron, :glutSolidTetrahedron, [], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method wire_icosahedron()
-  # @return [nil] 
+  # @return [nil]
   # @scope class
   attach_function :wire_icosahedron, :glutWireIcosahedron, [], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method solid_icosahedron()
-  # @return [nil] 
+  # @return [nil]
   # @scope class
   attach_function :solid_icosahedron, :glutSolidIcosahedron, [], :void
-  
+
   # Teapot rendering functions, found in freeglut_teapot.c
-  # 
+  #
   # @method wire_teapot(size)
-  # @param [Float] size 
-  # @return [nil] 
+  # @param [Float] size
+  # @return [nil]
   # @scope class
   attach_function :wire_teapot, :glutWireTeapot, [:double], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method solid_teapot(size)
-  # @param [Float] size 
-  # @return [nil] 
+  # @param [Float] size
+  # @return [nil]
   # @scope class
   attach_function :solid_teapot, :glutSolidTeapot, [:double], :void
-  
+
   # Game mode functions, see freeglut_gamemode.c
-  # 
+  #
   # @method game_mode_string(string)
-  # @param [String] string 
-  # @return [nil] 
+  # @param [String] string
+  # @return [nil]
   # @scope class
   attach_function :game_mode_string, :glutGameModeString, [:string], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method enter_game_mode()
-  # @return [Integer] 
+  # @return [Integer]
   # @scope class
   attach_function :enter_game_mode, :glutEnterGameMode, [], :int
-  
+
   # (Not documented)
-  # 
+  #
   # @method leave_game_mode()
-  # @return [nil] 
+  # @return [nil]
   # @scope class
   attach_function :leave_game_mode, :glutLeaveGameMode, [], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method game_mode_get(query)
-  # @param [Integer] query 
-  # @return [Integer] 
+  # @param [Integer] query
+  # @return [Integer]
   # @scope class
   attach_function :game_mode_get, :glutGameModeGet, [:uint], :int
-  
+
   # Video resize functions, see freeglut_videoresize.c
-  # 
+  #
   # @method video_resize_get(query)
-  # @param [Integer] query 
-  # @return [Integer] 
+  # @param [Integer] query
+  # @return [Integer]
   # @scope class
   attach_function :video_resize_get, :glutVideoResizeGet, [:uint], :int
-  
+
   # (Not documented)
-  # 
+  #
   # @method setup_video_resizing()
-  # @return [nil] 
+  # @return [nil]
   # @scope class
   attach_function :setup_video_resizing, :glutSetupVideoResizing, [], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method stop_video_resizing()
-  # @return [nil] 
+  # @return [nil]
   # @scope class
   attach_function :stop_video_resizing, :glutStopVideoResizing, [], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method video_resize(x, y, width, height)
-  # @param [Integer] x 
-  # @param [Integer] y 
-  # @param [Integer] width 
-  # @param [Integer] height 
-  # @return [nil] 
+  # @param [Integer] x
+  # @param [Integer] y
+  # @param [Integer] width
+  # @param [Integer] height
+  # @return [nil]
   # @scope class
   attach_function :video_resize, :glutVideoResize, [:int, :int, :int, :int], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method video_pan(x, y, width, height)
-  # @param [Integer] x 
-  # @param [Integer] y 
-  # @param [Integer] width 
-  # @param [Integer] height 
-  # @return [nil] 
+  # @param [Integer] x
+  # @param [Integer] y
+  # @param [Integer] width
+  # @param [Integer] height
+  # @return [nil]
   # @scope class
   attach_function :video_pan, :glutVideoPan, [:int, :int, :int, :int], :void
-  
+
   # Colormap functions, see freeglut_misc.c
-  # 
+  #
   # @method set_color(color, red, green, blue)
-  # @param [Integer] color 
-  # @param [Float] red 
-  # @param [Float] green 
-  # @param [Float] blue 
-  # @return [nil] 
+  # @param [Integer] color
+  # @param [Float] red
+  # @param [Float] green
+  # @param [Float] blue
+  # @return [nil]
   # @scope class
   attach_function :set_color, :glutSetColor, [:int, :float, :float, :float], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method get_color(color, component)
-  # @param [Integer] color 
-  # @param [Integer] component 
-  # @return [Float] 
+  # @param [Integer] color
+  # @param [Integer] component
+  # @return [Float]
   # @scope class
   attach_function :get_color, :glutGetColor, [:int, :int], :float
-  
+
   # (Not documented)
-  # 
+  #
   # @method copy_colormap(window)
-  # @param [Integer] window 
-  # @return [nil] 
+  # @param [Integer] window
+  # @return [nil]
   # @scope class
   attach_function :copy_colormap, :glutCopyColormap, [:int], :void
-  
+
   # Misc keyboard and joystick functions, see freeglut_misc.c
-  # 
+  #
   # @method ignore_key_repeat(ignore)
-  # @param [Integer] ignore 
-  # @return [nil] 
+  # @param [Integer] ignore
+  # @return [nil]
   # @scope class
   attach_function :ignore_key_repeat, :glutIgnoreKeyRepeat, [:int], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method set_key_repeat(repeat_mode)
-  # @param [Integer] repeat_mode 
-  # @return [nil] 
+  # @param [Integer] repeat_mode
+  # @return [nil]
   # @scope class
   attach_function :set_key_repeat, :glutSetKeyRepeat, [:int], :void
-  
+
   # (Not documented)
-  # 
+  #
   # @method force_joystick_func()
-  # @return [nil] 
+  # @return [nil]
   # @scope class
   attach_function :force_joystick_func, :glutForceJoystickFunc, [], :void
-  
+
   # Misc functions, see freeglut_misc.c
-  # 
+  #
   # @method extension_supported(extension)
-  # @param [String] extension 
-  # @return [Integer] 
+  # @param [String] extension
+  # @return [Integer]
   # @scope class
   attach_function :extension_supported, :glutExtensionSupported, [:string], :int
-  
+
   # (Not documented)
-  # 
+  #
   # @method report_errors()
-  # @return [nil] 
+  # @return [nil]
   # @scope class
   attach_function :report_errors, :glutReportErrors, [], :void
-  
+
 end
