+++ test/test_utils.rb
@@ -1,5 +1,5 @@
 require "fileutils"
-require "ffi_gen"
+require "ffi_generator"
 
 def find_headers(dir, prefix = "")
   Dir.chdir dir do
@@ -11,7 +11,7 @@ def run_test(options = {})
   options[:files].each do |header|
     output_file = "#{File.dirname(__FILE__)}/output/#{header.sub(/\.h$/, ".rb")}"
     FileUtils.mkdir_p File.dirname(output_file)
-    
+
     FFIGen.generate(
       module_name: options[:library_name],
       ffi_lib:     options[:ffi_lib],
@@ -22,9 +22,9 @@ def run_test(options = {})
       blocking:    options.fetch(:blocking, []),
       output:      output_file
     )
-    
+
     require output_file
   end
-  
+
   puts "#{options[:library_name]} test successful"
 end
